+++ battleservice/src/services/battle/scene/birth/IScene.go
@@ -3,5 +3,5 @@ package birth
 type IScene interface {
 	GetRandPos() (x, y float64)
 	GetEntityID() uint64
-	GenBallID() uint32
+	GenBallID() uint64
 }
